Rename mockChromeApi
